add checking correctness of password to use the menu

+++ Menu/UsunKat/deleteCategory.cpp
@@ -7,35 +7,37 @@
 #include <fstream>
 #include <unistd.h>
 
-void deleteCategory::deleteCat(std::vector<std::vector<std::string>>& vec) {
+void deleteCategory::deleteCat(std::vector<std::vector<std::string>> &vec, bool isGoodPassword) {
+    if (!isGoodPassword) std::cerr << "void deleteCat(): Blad! Opcja niedostepna przez niepoprawne haslo do programu" << std::endl;
+    else {
+        std::string nameOfCategory;
+        int index = 0;
+        nlohmann::json j;
+        std::ifstream i("../categories.json");
+        i >> j;
+        i.close();
 
-    std::string nameOfCategory;
-    int index = 0;
-    nlohmann::json j;
-    std::ifstream i("../categories.json");
-    i >> j;
-    i.close();
+        std::cout
+                << "Wpisz nazwe kategorii ktora chcesz usunac\n!UWAGA!\nUsuniecie kategorii wiaze sie z usunieciem wszystkich hasel wraz z wpisana kategoria: "
+                << std::endl;
+        std::cin >> nameOfCategory;
 
-    std::cout
-            << "Wpisz nazwe kategorii ktora chcesz usunac\n!UWAGA!\nUsuniecie kategorii wiaze sie z usunieciem wszystkich hasel wraz z wpisana kategoria: "
-            << std::endl;
-    std::cin >> nameOfCategory;
+        for (auto i: j["categories"]) {
+            if (decrypt(i.get<std::string>()) == nameOfCategory) {
+                deletePasswordsContainCurrentCategory(nameOfCategory, vec);
+                j["categories"].erase(index);
+                break;
+            } else index++;
+        }
 
-    for (auto i: j["categories"]) {
-        if (decrypt(i.get<std::string>()) == nameOfCategory) {
-            deletePasswordsContainCurrentCategory(nameOfCategory, vec);
-            j["categories"].erase(index);
-            break;
-        } else index++;
+        std::ofstream o("../categories.json");
+        o << std::setw(4) << j << std::endl;
+        o.close();
+
+        std::cout << "Kategoria " << nameOfCategory << " zostala usunieta!" << std::endl;
+
+        sleep(2);
     }
-
-    std::ofstream o("../categories.json");
-    o << std::setw(4) << j << std::endl;
-    o.close();
-
-    std::cout << "Kategoria " << nameOfCategory << " zostala usunieta!" << std::endl;
-
-    sleep(2);
 }
 
 void deleteCategory::deletePasswordsContainCurrentCategory(const std::string& nameOfCategory,
